+++ src/Either/Left.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Innmind\Immutable\Either;
 
-use Innmind\Immutable\Either;
+use Innmind\Immutable\{
+    Either,
+    Maybe,
+};
 
 /**
  * @template L1
@@ -61,5 +64,10 @@ final class Left implements Implementation
     public function filter(callable $predicate, callable $otherwise): self
     {
         return $this;
+    }
+
+    public function maybe(): Maybe
+    {
+        return Maybe::nothing();
     }
 }